Bundle Compiler refactoring pass
Previously, all of the bundle compiler implementation was in a single file called `templates.ts`. This commit breaks that file into logical units and also adds some additional inline documentation.
