+++ packages/@glimmer/bundle-compiler/index.ts
@@ -1,1 +1,20 @@
-export * from './lib/templates';
+export {
+  BundleCompiler,
+  BundleCompilerOptions,
+  BundleCompileOptions,
+  BundleCompilationResult
+} from './lib/bundle-compiler';
+
+export {
+  CompilerDelegate
+} from './lib/compiler-delegate';
+
+export {
+  Specifier,
+  specifierFor
+} from './lib/specifiers';
+
+export {
+  SpecifierMap,
+  LookupMap
+} from './lib/specifier-map';
